Fix issues after upgrading to micronaut 4.0.0-SNAPSHOT
