+++ docs-examples/example-java/src/test/java/io/micronaut/nats/AbstractNatsTest.java
@@ -1,13 +1,14 @@
 package io.micronaut.nats;
 
-import java.util.HashMap;
-import java.util.Map;
-import java.util.concurrent.Callable;
-
 import io.micronaut.context.ApplicationContext;
 import org.junit.jupiter.api.AfterEach;
 import org.testcontainers.containers.GenericContainer;
 import org.testcontainers.containers.wait.strategy.LogMessageWaitStrategy;
+
+import java.util.HashMap;
+import java.util.List;
+import java.util.Map;
+import java.util.concurrent.Callable;
 
 import static java.util.concurrent.TimeUnit.SECONDS;
 import static org.awaitility.Awaitility.await;
@@ -34,6 +35,8 @@ public abstract class AbstractNatsTest {
         config.put("nats.default.addresses",
             "nats://localhost:" + natsContainer.getMappedPort(4222));
         config.put("spec.name", getClass().getSimpleName());
+        config.put("nats.default.jetstream.streams.events.storage-type", "Memory");
+        config.put("nats.default.jetstream.streams.events.subjects", List.of("events.>"));
         return config;
     }
